feat: Add route for checking username uniqueness
The code changes include adding a new route for checking the uniqueness of a username. This route is responsible for validating the query parameters and checking if the username already exists and is verified. If the username is unique, a success message is returned. If the username already exists, an error message is returned.

Closes #456